example(): Update examples to use new options
