+++ examples/nest-graphql-sequelize/src/todo-item/todo-item.module.ts
@@ -1,13 +1,35 @@
+import { NestjsQueryGraphQLModule } from '@nestjs-query/query-graphql';
 import { NestjsQuerySequelizeModule } from '@nestjs-query/query-sequelize';
 import { Module } from '@nestjs/common';
+import { AuthGuard } from '../auth.guard';
+import { TodoItemInputDTO } from './dto/todo-item-input.dto';
+import { TodoItemUpdateDTO } from './dto/todo-item-update.dto';
+import { TodoItemDTO } from './dto/todo-item.dto';
 import { TodoItemEntityTags } from './entity/todo-item-tag.entity';
 import { TodoItemAssembler } from './todo-item.assembler';
 import { TodoItemEntity } from './entity/todo-item.entity';
 import { TodoItemResolver } from './todo-item.resolver';
-import { TodoItemService } from './todo-item.service';
+
+const guards = [AuthGuard];
 
 @Module({
-  providers: [TodoItemResolver, TodoItemService, TodoItemAssembler],
-  imports: [NestjsQuerySequelizeModule.forFeature([TodoItemEntity, TodoItemEntityTags])],
+  providers: [TodoItemResolver],
+  imports: [
+    NestjsQueryGraphQLModule.forFeature({
+      imports: [NestjsQuerySequelizeModule.forFeature([TodoItemEntity, TodoItemEntityTags])],
+      assemblers: [TodoItemAssembler],
+      resolvers: [
+        {
+          DTOClass: TodoItemDTO,
+          AssemblerClass: TodoItemAssembler,
+          CreateDTOClass: TodoItemInputDTO,
+          UpdateDTOClass: TodoItemUpdateDTO,
+          create: { guards },
+          update: { guards },
+          delete: { guards },
+        },
+      ],
+    }),
+  ],
 })
 export class TodoItemModule {}
