+++ examples/nest-graphql-typeorm-soft-delete/src/todo-item/todo-item.module.ts
@@ -1,11 +1,28 @@
+import { NestjsQueryGraphQLModule } from '@nestjs-query/query-graphql';
 import { NestjsQueryTypeOrmModule } from '@nestjs-query/query-typeorm';
 import { Module } from '@nestjs/common';
+import { TodoItemInputDTO } from './dto/todo-item-input.dto';
+import { TodoItemUpdateDTO } from './dto/todo-item-update.dto';
+import { TodoItemDTO } from './dto/todo-item.dto';
 import { TodoItemEntity } from './todo-item.entity';
 import { TodoItemResolver } from './todo-item.resolver';
 import { TodoItemService } from './todo-item.service';
 
 @Module({
-  providers: [TodoItemResolver, TodoItemService],
-  imports: [NestjsQueryTypeOrmModule.forFeature([TodoItemEntity])],
+  providers: [TodoItemResolver],
+  imports: [
+    NestjsQueryGraphQLModule.forFeature({
+      imports: [NestjsQueryTypeOrmModule.forFeature([TodoItemEntity])],
+      services: [TodoItemService],
+      resolvers: [
+        {
+          DTOClass: TodoItemDTO,
+          ServiceClass: TodoItemService,
+          CreateDTOClass: TodoItemInputDTO,
+          UpdateDTOClass: TodoItemUpdateDTO,
+        },
+      ],
+    }),
+  ],
 })
 export class TodoItemModule {}
