+++ examples/nest-graphql-typeorm-soft-delete/src/todo-item/todo-item.resolver.ts
@@ -1,19 +1,12 @@
 import { UpdateManyResponse, Filter } from '@nestjs-query/core';
-import { CRUDResolver, FilterType, UpdateManyResponseType } from '@nestjs-query/query-graphql';
+import { FilterType, UpdateManyResponseType } from '@nestjs-query/query-graphql';
 import { Resolver, Args, Mutation, ID } from '@nestjs/graphql';
-import { TodoItemInputDTO } from './dto/todo-item-input.dto';
-import { TodoItemUpdateDTO } from './dto/todo-item-update.dto';
 import { TodoItemDTO } from './dto/todo-item.dto';
 import { TodoItemService } from './todo-item.service';
 
 @Resolver(() => TodoItemDTO)
-export class TodoItemResolver extends CRUDResolver(TodoItemDTO, {
-  CreateDTOClass: TodoItemInputDTO,
-  UpdateDTOClass: TodoItemUpdateDTO,
-}) {
-  constructor(readonly service: TodoItemService) {
-    super(service);
-  }
+export class TodoItemResolver {
+  constructor(readonly service: TodoItemService) {}
 
   @Mutation(() => TodoItemDTO)
   restoreOneTodoItem(@Args('input', { type: () => ID }) id: number): Promise<TodoItemDTO> {
