+++ examples/nest-graphql-typeorm/src/todo-item/todo-item.module.ts
@@ -1,12 +1,33 @@
+import { NestjsQueryGraphQLModule } from '@nestjs-query/query-graphql';
 import { NestjsQueryTypeOrmModule } from '@nestjs-query/query-typeorm';
 import { Module } from '@nestjs/common';
+import { AuthGuard } from '../auth.guard';
+import { TodoItemInputDTO } from './dto/todo-item-input.dto';
+import { TodoItemUpdateDTO } from './dto/todo-item-update.dto';
+import { TodoItemDTO } from './dto/todo-item.dto';
 import { TodoItemAssembler } from './todo-item.assembler';
 import { TodoItemEntity } from './todo-item.entity';
 import { TodoItemResolver } from './todo-item.resolver';
-import { TodoItemService } from './todo-item.service';
 
+const guards = [AuthGuard];
 @Module({
-  providers: [TodoItemResolver, TodoItemService, TodoItemAssembler],
-  imports: [NestjsQueryTypeOrmModule.forFeature([TodoItemEntity])],
+  providers: [TodoItemResolver],
+  imports: [
+    NestjsQueryGraphQLModule.forFeature({
+      imports: [NestjsQueryTypeOrmModule.forFeature([TodoItemEntity])],
+      assemblers: [TodoItemAssembler],
+      resolvers: [
+        {
+          DTOClass: TodoItemDTO,
+          AssemblerClass: TodoItemAssembler,
+          CreateDTOClass: TodoItemInputDTO,
+          UpdateDTOClass: TodoItemUpdateDTO,
+          create: { guards },
+          update: { guards },
+          delete: { guards },
+        },
+      ],
+    }),
+  ],
 })
 export class TodoItemModule {}
