+++ jest.e2e.config.js
@@ -5,4 +5,8 @@ module.exports = {
   // from the above list
   preset: 'ts-jest',
   testMatch: ['**/e2e/**/*.spec.ts'],
+  collectCoverageFrom: ['packages/**/*.ts', '!**/__tests__/**', '!**/dist/**', '!**/node_modules/**'],
+  moduleNameMapper: {
+    '^@nestjs-query/(.*)$': '<rootDir>/packages/$1/src',
+  },
 };
